--- Balls Coming/Assets/_Project/Scripts/Database.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b747abd02c9f86341b821f32aab2dabe
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
