--- Balls Coming/Assets/_Project/Scripts/Database/LeaderboardsManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ef66e08e07077064e892a39f1c7064a4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
