--- Balls Coming/Assets/_Project/Scripts/Database/PlayerLogin.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a40734b300b37394d9bc549972a9f928
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
