+++ Balls Coming/Assets/_Project/Scripts/Server Side/ServerSide.asmdef
@@ -1,0 +1,17 @@
+{
+    "name": "ServerSide",
+    "rootNamespace": "",
+    "references": [
+        "GUID:6055be8ebefd69e48b49212b09b47b2f",
+        "GUID:dffdc6d84e1ec4e4f867437aa21264d5"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
