+++ Balls Coming/Assets/_Project/Scripts/UI/CoinCollection.cs
@@ -1,3 +1,5 @@
+using BallsComing.ServerSide;
+
 using UnityEngine;
 using UnityEngine.Events;
 using TMPro;
