+++ Balls Coming/Packages/manifest.json
@@ -1,10 +1,11 @@
 {
   "dependencies": {
     "com.lootlocker.lootlockersdk": "https://github.com/LootLocker/unity-sdk.git",
-    "com.unity.collab-proxy": "2.0.1",
+    "com.unity.collab-proxy": "2.0.3",
+    "com.unity.connect.share": "4.2.3",
     "com.unity.feature.development": "1.0.1",
-    "com.unity.ide.rider": "3.0.18",
-    "com.unity.ide.visualstudio": "2.0.17",
+    "com.unity.ide.rider": "3.0.20",
+    "com.unity.ide.visualstudio": "2.0.18",
     "com.unity.ide.vscode": "1.2.5",
     "com.unity.test-framework": "1.1.31",
     "com.unity.textmeshpro": "3.0.6",
